UPDATE: Formulario de registro de recuentos, se actualiza el icono de cerrar modal y estilos del formulario
