+++ src/features/pRecuentos/components/registroRP/RegistroRecuento.module.css
@@ -1,0 +1,12 @@
+.buttonContainer{
+    display: flex;
+    justify-content: center;
+    gap: 20px;
+    margin-top: 20px;
+}
+
+.inputContainer {
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+}
