+++ src/shared/components/modals/modal.module.css
@@ -10,10 +10,10 @@
   
   .modal {
     background: #fff;
-    padding: 20px;
+    padding: 30px;
     border-radius: 12px;
-    width: 90%;
-    max-width: 500px;
+    width: 30%;
+    max-width: 50%;
     box-shadow: 0 4px 20px rgba(0,0,0,0.2);
   }
   
@@ -22,6 +22,13 @@
     justify-content: space-between;
     align-items: center;
     margin-bottom: 10px;
+  }
+
+  .header h2 {
+    font-size: 2rem;
+    color: var(--color-secundario);
+    text-align: center;
+    font-weight: 700;
   }
   
   .closeButton {
